move data.py to services
